Split up processing from enqueue code
